Move list files function into common crate
